update CETEIcean to v1.2.0+ (6/15/2020) (which includes one small change I asked for and removal of the `table` behaviour) *and* my pull request https://github.com/TEIC/CETEIcean/pull/35: - style TEI tables using `display` values `table`, `table-row`, `table-cell`; - added `.js` extension to the imported module names;
my custom `tei.scss`:
- lower-cased TEI element names;
- style TEI tables as HTML tables are styled by `minima`;

- semicolon cleanup;
- do not use removed addStyle();

- names aligned with CETEIcean;
