+++ docs/_sass/main.scss
@@ -1,5 +1,5 @@
 @import "minima";
 @import "blog";
 @import "facsimile";
-@import "cetei";
+@import "CETEIcean";
 @import "tei";
