+++ docs/js/CETEI/CETEI.js
@@ -1,22 +1,22 @@
-// LMD added '.js' in all the includes.
-import * as utilities from './utilities.js';
 import defaultBehaviors from './defaultBehaviors.js';
-import {addBehaviors, addBehavior, applyBehaviors} from './behaviors.js';
+import * as utilities from './utilities.js';
+import {addBehaviors, addBehavior, applyBehaviors, removeBehavior} from './behaviors.js';
 import {learnElementNames} from './dom.js';
 
 class CETEI {
   constructor(options){
-    this.options = options ? options : {}
+    this.options = options ? options : {};
 
     // Bind methods
     this.addBehaviors = addBehaviors.bind(this);
     this.addBehavior = addBehavior.bind(this);
     this.applyBehaviors = applyBehaviors.bind(this);
+    this.removeBehavior = removeBehavior.bind(this);
 
     // Bind selected utilities
-    this.utilities = {}
+    this.utilities = {};
     for (const u of Object.keys(utilities)) {
-      if (u == "rw") {
+      if (["getPrefixDef", "rw", "resolveURI"].includes(u)) {
         this.utilities[u] = utilities[u].bind(this);
       } else {
         this.utilities[u] = utilities[u];
@@ -30,6 +30,7 @@ class CETEI {
     this.hasStyle = false;
     this.prefixDefs = [];
     this.debug = this.options.debug === true ? true : false;
+
     if (this.options.base) {
       this.base = this.options.base;
     } else {
@@ -43,6 +44,11 @@ class CETEI {
     }
     if (!this.options.omitDefaultBehaviors) {
       this.addBehaviors(defaultBehaviors);
+    }
+    if (this.options.ignoreFragmentId) {
+      if (window) {
+        window.removeEventListener("ceteiceanload", CETEI.restorePosition);
+      }
     }
   }
 
@@ -182,7 +188,7 @@ class CETEI {
         perElementFn(newElement, el);
       }
       return newElement;
-    }
+    };
 
     this.dom = convertEl(XML_dom.documentElement);
 
@@ -198,16 +204,6 @@ class CETEI {
         window.dispatchEvent(ceteiceanLoad);
       }
       return this.dom;
-    }
-  }
-
-  /* 
-    If the TEI document defines CSS styles in its tagsDecl, this method
-    copies them into the wrapper HTML document's head.
-  */
-  addStyle(doc, data) {
-    if (this.hasStyle) {
-      doc.getElementsByTagName("head").item(0).appendChild(data.getElementsByTagName("style").item(0).cloneNode(true));
     }
   }
 
@@ -229,9 +225,263 @@ class CETEI {
     this.base = base;
   }
 
+  /*
+  Appends any element returned by the function passed in the first
+  parameter to the element in the second parameter. If the function
+  returns nothing, this is a no-op aside from any side effects caused
+  by the provided function.
+
+  Called by getHandler() and fallback()
+*/
+append(fn, elt) {
+  let self = this;
+  if (elt) {
+    let content = fn.call(self.utilities, elt);
+    if (content && !self.childExists(elt.firstElementChild, content.nodeName)) {
+      self.appendBasic(elt, content);
+    }
+  } else {
+    return function() {
+      if (!this.hasAttribute("data-processed")) {
+        let content = fn.call(self.utilities, this);
+        if (content && !self.childExists(this.firstElementChild, content.nodeName)) {
+          self.appendBasic(this, content);
+        }
+      }
+    }
+  }
+}
+
+appendBasic(elt, content) {
+  utilities.hideContent(elt);
+  elt.appendChild(content);
+}
+
+// Given an element, return its qualified name as defined in a behaviors object
+bName(e) {
+  return e.tagName.substring(0,e.tagName.indexOf("-")).toLowerCase() + ":" + e.getAttribute("data-origname");
+}
+
+/*
+  Private method called by append(). Takes a child element and a name, and recurses through the
+  child's siblings until an element with that name is found, returning true if it is and false if not.
+*/
+childExists(elt, name) {
+  if (elt && elt.nodeName == name) {
+    return true;
+  } else {
+    return elt && elt.nextElementSibling && this.childExists(elt.nextElementSibling, name);
+  }
+}
+
+/*
+  Takes a template in the form of either an array of 1 or 2
+  strings or an object with CSS selector keys and either functions
+  or arrays as described above. Returns a closure around a function
+  that can be called in the element constructor or applied to an
+  individual element.
+
+  Called by the getHandler() and getFallback() methods
+*/
+decorator(template) {
+  if (Array.isArray(template) && !Array.isArray(template[0])) {
+    return this.applyDecorator(template)
+  }
+  let self = this;
+  return function(elt) {
+    for (let rule of template) {
+      if (elt.matches(rule[0]) || rule[0] === "_") {
+        if (Array.isArray(rule[1])) {
+          return self.decorator(rule[1]).call(this, elt);
+        } else {
+          return rule[1].call(this, elt);
+        }
+      }
+    }
+  }
+}
+
+applyDecorator(strings) {
+  let self = this;
+  return function (elt) {
+    let copy = [];
+    for (let i = 0; i < strings.length; i++) {
+      copy.push(self.template(strings[i], elt));
+    }
+    return self.insert(elt, copy);
+  }
+}
+
+/*
+  Returns the fallback function for the given element name.
+  Called by fallback().
+*/
+getFallback(behaviors, fn) {
+  if (behaviors[fn]) {
+    if (behaviors[fn] instanceof Function) {
+      return behaviors[fn];
+    } else {
+      return decorator(behaviors[fn]);
+    }
+  }
+}
+
+/*
+  Returns the handler function for the given element name
+  Called by define().
+*/
+getHandler(behaviors, fn) {
+  if (behaviors[fn]) {
+    if (behaviors[fn] instanceof Function) {
+      return this.append(behaviors[fn]);
+    } else {
+      return this.append(this.decorator(behaviors[fn]));
+    }
+  }
+}
+
+insert(elt, strings) {
+  let span = document.createElement("span");
+  for (let node of Array.from(elt.childNodes)) {
+    if (node.nodeType === Node.ELEMENT_NODE && !node.hasAttribute("data-processed")) {
+      this.processElement(node);
+    }
+  }
+  // If we have before and after tags have them parsed by
+  // .innerHTML and then add the content to the resulting child
+  if (strings[0].match("<[^>]+>") && strings[1] && strings[1].match("<[^>]+>")) {
+    span.innerHTML = strings[0] + (strings[1]?strings[1]:"");
+    for (let node of Array.from(elt.childNodes)) {
+      span.firstElementChild.appendChild(node.cloneNode(true));
+    }
+  } else {
+    span.innerHTML = strings[0];
+    span.setAttribute("data-before", strings[0].replace(/<[^>]+>/g,"").length);
+    for (let node of Array.from(elt.childNodes)) {
+      span.appendChild(node.cloneNode(true));
+    }
+    if (strings.length > 1) {
+      span.innerHTML += strings[1];
+      span.setAttribute("data-after", strings[1].replace(/<[^>]+>/g,"").length);
+    }
+  }
+  return span;
+}
+
+// Runs behaviors recursively on the supplied element and children
+processElement(elt) {
+  if (elt.hasAttribute("data-origname") && ! elt.hasAttribute("data-processed")) {
+    let fn = this.getFallback(this.bName(elt));
+    if (fn) {
+      this.append(fn,elt);
+      elt.setAttribute("data-processed", "");
+    }
+  }
+  for (let node of Array.from(elt.childNodes)) {
+    if (node.nodeType === Node.ELEMENT_NODE) {
+      this.processElement(node);
+    }
+  }
+}
+
+// Given a qualified name (e.g. tei:text), return the element name
+tagName(name) {
+  if (name.includes(":"), 1) {
+    return name.replace(/:/,"-").toLowerCase();
+  } else {
+    return "ceteicean-" + name.toLowerCase();
+  }
+}
+
+template(str, elt) {
+  let result = str;
+  if (str.search(/\$(\w*)(@([a-zA-Z:]+))/ )) {
+    let re = /\$(\w*)@([a-zA-Z:]+)/g;
+    let replacements;
+    while (replacements = re.exec(str)) {
+      if (elt.hasAttribute(replacements[2])) {
+        if (replacements[1] && utilities[replacements[1]]) {
+          result = result.replace(replacements[0], utilities[replacements[1]](elt.getAttribute(replacements[2])));
+        } else {
+          result = result.replace(replacements[0], elt.getAttribute(replacements[2]));
+        }
+      } else {
+        result = result.replace(replacements[0], "");
+      }
+    }
+  }
+  return result;
+}
+
+/*
+  Registers the list of elements provided with the browser.
+  Called by makeHTML5(), but can be called independently if, for example,
+  you've created Custom Elements via an XSLT transformation instead.
+*/
+define(names) {
+  for (let name of names) {
+    try {
+      let fn = this.getHandler(this.behaviors, name);
+      window.customElements.define(this.tagName(name), class extends HTMLElement {
+        constructor() {
+          super();
+          if (!this.matches(":defined")) { // "Upgraded" undefined elements can have attributes & children; new elements can't
+            if (fn) {
+              fn.call(this);
+            }
+            // We don't want to double-process elements, so add a flag
+            this.setAttribute("data-processed", "");
+          }
+        }
+        // Process new elements when they are connected to the browser DOM
+        connectedCallback() {
+          if (!this.hasAttribute("data-processed")) {
+            if (fn) {
+              fn.call(this);
+            }
+            this.setAttribute("data-processed", "");
+          }
+        };
+      });
+    } catch (error) {
+      // When using the same CETEIcean instance for multiple TEI files, this error becomes very common.
+      // It's muted by default unless the debug option is set.
+      if (this.debug) {
+          console.log(tagName(name) + " couldn't be registered or is already registered.");
+          console.log(error);
+      }
+    }
+  }
+}
+
+/*
+  Provides fallback functionality for browsers where Custom Elements
+  are not supported.
+
+  Like define(), this is called by makeHTML5(), but can be called
+  independently.
+*/
+fallback(names) {
+  for (let name of names) {
+    let fn = getFallback(this.behaviors, name);
+    if (fn) {
+      for (let elt of Array.from((
+          this.dom && !this.done
+          ? this.dom
+          : document
+        ).getElementsByTagName(tagName(name)))) {
+        if (!elt.hasAttribute("data-processed")) {
+          append(fn, elt);
+        }
+      }
+    }
+  }
+}
+
   /**********************
    * Utility functions  *
    **********************/
+
 
   static savePosition() {
     window.sessionStorage.setItem(window.location + "-scroll", window.scrollY);
@@ -247,7 +497,6 @@ class CETEI {
       }
     } else {
       setTimeout(function() {
-        // LMD added decodeURI() to handle Unicode ids:
         let h = document.querySelector(window.decodeURI(window.location.hash));
         if (h) {
           h.scrollIntoView();
@@ -268,5 +517,5 @@ try {
 } catch (e) {
   console.log(e);
 }
-// Make main class available to pre-ES6 browser environments
-export default CETEI;
+
+export default CETEI
