+++ docs/js/CETEI/dom.js
@@ -1,18 +1,16 @@
 export function learnElementNames(XML_dom, namespaces) {
   const root = XML_dom.documentElement;
+  let i = 1;
+  let qname = function(e) {
+    if (!namespaces.has(e.namespaceURI ? e.namespaceURI : "")) {
+      namespaces.set(e.namespaceURI, "ns" + i++);
+    }
+    return namespaces.get(e.namespaceURI ? e.namespaceURI : "") + ":" + e.localName;
+  };
   const els = new Set(
-    Array.from(root.querySelectorAll("*"),
-    e => (
-      namespaces.has(e.namespaceURI ? e.namespaceURI : "")
-      ? namespaces.get(e.namespaceURI ? e.namespaceURI : "") + ":"
-      : ""
-    ) + e.localName) );
+    Array.from(root.querySelectorAll("*"), qname));
+
   // Add the root element to the array
-  els.add(
-    (
-      namespaces.has(root.namespaceURI ? root.namespaceURI : "")
-      ? namespaces.get(root.namespaceURI ? root.namespaceURI : "") + ":"
-      : ""
-    ) + root.localName);
+  els.add(qname(root));
   return els
 }
